SettingsPage: Add checkboxes for setting the save username/password options
Currently, the username/password settings are wiped whenever the user
toggles the save username/password settings, which might be a little
unintuitive. This can be revisited if we set up a hook to save settings
immediately before exit, where we can just erase the login credentials
before saving, or maybe using a custom Serde serializer to avoid
serializing these properties when the `save_*` property is false.
called,